Refactor visibility of assertion internals
